Fix FunctionCallTag registering in iOS and support to manually refreshing the container (#316)
This PR fixes an error in `FunctionCallTagHandlerIOS.ts` module which caused the app to to crash due to a method call on an uninitialized variable.

It also adds `refreshContainer` method so that developers can control the refresh rate of the contianer. @cbrevik, I did not document this method in the README file because I was not able to see it working properly. The native module was acting up and always throwing too many request error, and the new variables that we defined in the GTM console never became available, but the bridge works as expected, since these logs come form the native module not the bridge itself. If you decide to document it in the README, make note of that (per Google's own words) this method should not be called more than once per each 15 minutes.